// mon-premier-projet/src/app/enum.ts
export enum SideNav {
    CALENDRIER,
    CATALOGUE_DE_FORMATION,
    FORMATEUR,
    STAGIAIRE,
}
export enum CatalogueFormation {
    DELL,
    FORTINET,
    MICROSOFT,
    VERITAS,
    VMWARE,
}
export enum TypeFormateurString {
    INTERNE = 'interne_',
    EXTERNE = 'externe_',
}

// export enum 
export enum TypeFormateur {
    INTERNE ,
    EXTERNE ,
}
export enum CoursDELL {
    dellEmcAvamarManagement = 'DELL EMC Avamar Management',
    dellEmcDataDomainSystemeAdminitration = 'DELL EMC Data Domain System Administration',
    dellEmcM1000eInstallationAdministationTrousbleshooting = 'DELL EMC M1000eInstallation, Administration and Troubleshooting',
    dellEmcNetworkerImplementationManagement = 'DELL EMC Networker Implementation and Management',
    dellEmcRecoverPoint = 'DELL EMC RecoverPoint',
    dellEmcScSeriesImplenetationAdministration = 'DELL EMC SC Series Implementation and Administration',
    dellEmcUnityImplementationADMINISTRATION =  'DELL EMC Unity Implementation and Administration',
    dellEmcVxrailApplianceAdministrationManagement = 'DELL EMC VxRail Appliance Administration and Management' ,
    dellNetworkingCampus = 'DELL Networking Campus' ,
    dellNetworkingDataCenterAdvancedFeaturesAdministration = 'DELL Networking Data Center Advanced Features & Administration' ,
}
export enum CoursFortinet {
    fortiAdc = 'FortiADC',
    fortiGateInfrastructure = 'FortiGate Infrastructure',
    fortiGateSecutity = 'FortiGate Security',
    fortiWeb = 'FortiWeb' ,
}
export enum CoursMicrosoft {
    AdministrationMicrosoftExchangeServer = 'Administration de Microsoft Exchange Server',
    IdentityWindowsServer2016 = 'Identity with Windows Server 2016',
    ImplementingSoftwareDefinedDatacenter = 'Implementing a Software-Defined Datacenter',
    InstallationStorageComputeWindowsServer2016 = 'Installation, Storage, and Compute with Windows Server 2016',
    Office365AdministrationTroubleshooting = 'Office 365 Administration and Troubleshooting',
    SQLServerAdministeringSQLDatabaseInfrastructure = 'SQL Server Administering a SQL Database Infrastructure',
    SystemCenterConfigurationManager = 'System Center Configuration Manager',
    SystemCenterOperationManager = 'System Center Operation Manager',
    SystemCenterServiceManager = 'System Center Service Manager',
}
export enum coursVeritas {
    VeritasBackupExecAdministration = 'Veritas Backup Exec Administration',
    VeritasNetBackup = 'Veritas NetBackup',
}
export enum CoursVmware {
    VMwareNSXInstallConfigureManage = 'VMware NSX Install, Configure, Manage',
    VMwarevCenterSiteRecoveryManager =  'VMware vCenter Site Recovery Manager',
    VMwarevRealizeOperations =  'VMware vRealize Operations',
    VMwarevSANdeployandmanage = 'VMware vSAN deploy and manage',
    VMwarevSphereInstallConfigureandManage = 'VMware vSphere Install, Configure and Manage',
    VMwarevSpheretroubleshootingworkshop = 'VMware vSphere troubleshooting workshop',
}